tests/wpt/tests/html/semantics/embedded-content/media-elements/track/track-element/track-load-from-src-readyState.html

<!DOCTYPE html>
<title>Load event on HTMLTrackElement and LOADED readyState on TextTrack when src is set from JavaScript</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<video>
    <track>
</video>
<script>
async_test(function(t) {
    var track = document.querySelector("track");
    assert_equals(track.readyState, HTMLTrackElement.NONE);

    track.onload = t.step_func_done(function() {
        assert_equals(track.readyState, HTMLTrackElement.LOADED);
    });

    track.src = "resources/webvtt-file.vtt";
    track.track.mode = "hidden";
});
</script>
